+++ examples/react-antd/src/antd-layout.tsx
@@ -2,10 +2,12 @@ import React from 'react';
 import {
   LaptopOutlined,
   NotificationOutlined,
-  UserOutlined,
+  UserOutlined
 } from '@ant-design/icons';
 import type { MenuProps } from 'antd';
 import { Breadcrumb, Layout, Menu, theme } from 'antd';
+
+import { useNavigate, Outlet } from 'react-router-dom';
 
 import './main.css';
 
@@ -13,13 +15,13 @@ const { Header, Content, Sider } = Layout;
 
 const items1: MenuProps['items'] = ['1', '2', '3'].map((key) => ({
   key,
-  label: `nav ${key}`,
+  label: `nav ${key}`
 }));
 
 const items2: MenuProps['items'] = [
   UserOutlined,
   LaptopOutlined,
-  NotificationOutlined,
+  NotificationOutlined
 ].map((icon, index) => {
   const key = String(index + 1);
 
@@ -32,16 +34,17 @@ const items2: MenuProps['items'] = [
       const subKey = index * 4 + j + 1;
       return {
         key: subKey,
-        label: `option${subKey}`,
+        label: `option${subKey}`
       };
-    }),
+    })
   };
 });
 
 export const AntdLayout: React.FC = () => {
   const {
-    token: { colorBgContainer },
+    token: { colorBgContainer }
   } = theme.useToken();
+  const navigate = useNavigate();
 
   return (
     <Layout>
@@ -52,6 +55,7 @@ export const AntdLayout: React.FC = () => {
           mode="horizontal"
           defaultSelectedKeys={['2']}
           items={items1}
+          onClick={({ key }) => navigate(`/${key}`)}
         />
       </Header>
       <Layout>
@@ -75,10 +79,10 @@ export const AntdLayout: React.FC = () => {
               padding: 24,
               margin: 0,
               minHeight: 280,
-              background: colorBgContainer,
+              background: colorBgContainer
             }}
           >
-            Content
+            <Outlet />
           </Content>
         </Layout>
       </Layout>